+++ src/SimpleRequest.php
@@ -5,6 +5,8 @@ namespace Mediawiki\Api;
 use InvalidArgumentException;
 
 /**
+ * Please consider using a FluentRequest object
+ *
  * @since 0.2
  */
 class SimpleRequest implements Request {
@@ -25,36 +27,46 @@ class SimpleRequest implements Request {
 	private $headers;
 
 	/**
+	 * @var RequestOptions
+	 */
+	private $options;
+
+	/**
 	 * @param string $action
 	 * @param array $params
 	 * @param array $headers
+	 * @param RequestOptions $options
 	 *
 	 * @throws InvalidArgumentException
 	 */
-	public function __construct( $action, array $params = array(), array $headers = array() ) {
+	public function __construct(
+		$action,
+		array $params = array(),
+		array $headers = array(),
+		RequestOptions $options = null
+	) {
 		if( !is_string( $action ) ) {
 			throw new InvalidArgumentException( '$action must be string' );
+		}
+		if( is_null( $options ) ) {
+			$options = new RequestOptions();
 		}
 		$this->action = $action;
 		$this->params = $params;
 		$this->headers = $headers;
+		$this->options = $options;
 	}
 
-	/**
-	 * @return array
-	 *
-	 * @since 0.2
-	 */
 	public function getParams() {
 		return array_merge( array( 'action' => $this->action ) , $this->params );
 	}
 
-	/**
-	 * @return array
-	 *
-	 * @since 0.3
-	 */
 	public function getHeaders() {
 		return $this->headers;
 	}
+
+	public function getOptions() {
+		return $this->options;
+	}
+
 }